# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Blog API",
        "description": "Documentation de l’API du projet Blog",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Serveur principal"
        }
    ],
    "paths": {
        "/api/articles": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Lister tous les articles",
                "description": "Récupère tous les articles ou ceux d'une catégorie spécifique si `categorie_id` est fourni.",
                "operationId": "getArticles",
                "parameters": [
                    {
                        "name": "categorie_id",
                        "in": "query",
                        "description": "Filtrer les articles par ID de catégorie",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 2
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des articles récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Article"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Aucun article trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Aucun article trouvé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/articles/{id}": {
            "delete": {
                "tags": [
                    "Articles"
                ],
                "summary": "Supprimer une article",
                "description": "Supprime une article existante à partir de son ID.",
                "operationId": "deleteArticle",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'article à supprimer",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Article supprimée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Article supprimée avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Article introuvable",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Article introuvable"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Article": {
                "title": "Article Resource",
                "description": "Structure d'un article retourné par l'API",
                "properties": {
                    "id": {
                        "description": "Identifiant unique de l'article",
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "description": "Titre de l'article",
                        "type": "string",
                        "example": "Introduction à Laravel"
                    },
                    "description": {
                        "description": "Contenu de l'article",
                        "type": "string",
                        "example": "Laravel est un framework PHP moderne..."
                    },
                    "image": {
                        "description": "URL complète de l'image si disponible",
                        "type": "string",
                        "example": "http://127.0.0.1:8000/storage/images/article1.jpg",
                        "nullable": true
                    },
                    "keywords": {
                        "description": "Mots-clés liés à l'article",
                        "type": "string",
                        "example": "laravel, php, framework",
                        "nullable": true
                    },
                    "categorie": {
                        "description": "Nom de la catégorie associée",
                        "type": "string",
                        "example": "Programmation"
                    },
                    "created_at": {
                        "description": "Date de création brute",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-10-23T09:30:00Z"
                    },
                    "created_at_formatted": {
                        "description": "Date de création formatée",
                        "type": "string",
                        "example": "2025-10-23 09:30:00"
                    },
                    "updated_at": {
                        "description": "Date de mise à jour brute",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-10-23T10:00:00Z"
                    },
                    "updated_at_formatted": {
                        "description": "Date de mise à jour formatée",
                        "type": "string",
                        "example": "2025-10-23 10:00:00"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}